refactor: extract monolithic server.js into modular src/ structure
Split the 2200-line server.js into controllers, routes, services,
middleware, config, and utils following the pattern established by
the existing auth module extraction. Also adds weekly auto-merge
to the release workflow and fixes validatePassword import bug in
authController.
