+++ src/middleware/auth.js
@@ -1,0 +1,10 @@
+function requireAuth(req, res, next) {
+  if (!req.session || !req.session.userId) {
+    return res.status(401).json({});
+  }
+  next();
+}
+
+module.exports = {
+  requireAuth
+};
