+++ src/routes/laden.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const ladenController = require('../controllers/ladenController');
+const { requireAuth } = require('../middleware/auth');
+
+router.get('/', requireAuth, ladenController.getAll);
+
+module.exports = router;
